feat: II cross domains support - single call to action
