+++ src/frontend/src/lib/i18n/en.json
@@ -224,7 +224,6 @@
 		"quote_8": "Galaxies within grasp forever.",
 		"quote_9": "Navigate stars, chase dreams.",
 		"quote_10": "Universe: endless exploration awaits.",
-		"future": "Sign in to build your Web3 future.",
 		"internet_identity": "Continue with Internet Identity"
 	},
 	"satellites": {
